server/requirements.txt
alembic==1.13.3
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.5.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
openai==1.47.1
pkgutil-resolve-name==1.3.10
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.4
zipp==3.20.2
